Beta

+++ index.php
@@ -14,53 +14,82 @@ if($table_exists == FALSE) {
 if(isset($_POST['clear'])) {
 	$filename = $_POST['filename'];
 	
-	$clear_participants = mysqli_query($connect, "TRUNCATE TABLE ".$filename);
-	$reset_participants = mysqli_query($connect, "ALTER TABLE ".$filename." SET AUTO_INCREMENT = 1");
-	
-	$clear_rounds = mysqli_query($connect, "TRUNCATE TABLE ".$filename."_rounds");
-	$reset_rounds = mysqli_query($connect, "ALTER TABLE ".$filename."_rounds SET AUTO_INCREMENT = 1");
-	
-	$clear_results = mysqli_query($connect, "DELETE FROM tournament_results WHERE filename = '".$filename."'");
-	
-	$succes = "<div id='succes'>De gegevens van het toernooi zijn verwijderd</div>";
+	do {
+		// Clear the participants table and reset auto_increment to 1
+		$clear_participants = mysqli_query($connect, "TRUNCATE TABLE ".$filename);
+		
+		if(!$clear_participants) {
+			$response = log_mysql_error(mysqli_error($connect), __LINE__, __FILE__, true);
+			
+			break;
+		}		
+		
+		// Clear the played tournament rounds table and reset auto_increment to 1
+		$clear_rounds = mysqli_query($connect, "TRUNCATE TABLE ".$filename."_rounds");
+		
+		if(!$clear_rounds) {
+			$response = log_mysql_error(mysqli_error($connect), __LINE__, __FILE__, true);
+			
+			break;
+		}
+		
+		// Delete the results of the tournament, tournament was never played
+		$clear_results = mysqli_query($connect, "DELETE FROM tournament_results WHERE filename = '".$filename."'");
+		
+		if(!$clear_results) {
+			$response = log_mysql_error(mysqli_error($connect), __LINE__, __FILE__, true);
+			
+			break;
+		}
+		
+		// Everything went right
+		$response = "<div id='succes'>De gegevens van het toernooi zijn verwijderd</div>";
+	} while(0);
 }
 
 
 // Delete the entire tournament and its' results
 if(isset($_POST['delete'])) {
 	$filename = $_POST['filename'];
-
-	$drop_participants = mysqli_query($connect, "DROP TABLE ".$filename);
 	
-	$drop_rounds = mysqli_query($connect, "DROP TABLE ".$filename."_rounds");
-	
-	$clear_results = mysqli_query($connect, "DELETE FROM tournament_results WHERE filename = '".$filename."'");
-	
-	unlink("tournaments/".$filename.".php");
-	
-	$succes = "<div id='succes'>Het toernooi is verwijderd</div>";
-}
-
-
-// Add a new tournament to the system
-if(isset($_POST['add']) && !empty($_POST['name'])) {
-	$year = $_POST['year'];
-	$name = htmlentities($_POST['name'], ENT_QUOTES, "UTF-8");
-	
-	if(!file_exists("tournaments/".$name."_".$year.".php")) {	
-		$file = "tournaments/tournament_year.php";
-		$newfile = "tournaments/".$name."_".$year.".php";
+	do {
+		// Delete the entire participants table
+		$drop_participants = mysqli_query($connect, "DROP TABLE ".$filename);
 		
-		if(!copy($file, $newfile)) {
-			$error = "<div id='error'>Er is iets mis gegaan, probeer het opnieuw</div>";
-		} else {
-			$succes = "<div id='succes'>Het toernooi is aangemaakt</div>";
+		if(!$drop_participants) {
+			$response = log_mysql_error(mysqli_error($connect), __LINE__, __FILE__, true);
+			
+			break;
 		}
-	} else {
-		$error = "<div id='error'>Deze editie bestaat al. Deze dient eerst verwijdert te worden.</div>";
-	}
-} elseif(isset($_POST['add']) && empty($_POST['name'])) {
-	$error = "<div id='error'>U heeft geen naam ingevuld</div>";
+		
+		// Delete the entire played tournament rounds table
+		$drop_rounds = mysqli_query($connect, "DROP TABLE ".$filename."_rounds");
+		
+		if(!$drop_rounds) {
+			$response = log_mysql_error(mysqli_error($connect), __LINE__, __FILE__, true);
+			
+			break;
+		}
+		
+		// Delete the results of the tournament
+		$clear_results = mysqli_query($connect, "DELETE FROM tournament_results WHERE filename = '".$filename."'");
+		
+		if(!$clear_results) {
+			$response = log_mysql_error(mysqli_error($connect), __LINE__, __FILE__, true);
+			
+			break;
+		}
+		
+		// Delete the file of the tournament
+		if(!unlink("tournaments/".$filename.".php")) {
+			$response = "<div id='error'>Er is iets mis gegaan, probeer het opnieuw</div>\n";
+			
+			break;
+		} 
+		
+		// Everything went right
+		$response = "<div id='succes'>Het toernooi is verwijderd</div>";
+	} while(0);
 }
 ?>
 <!DOCTYPE html>
@@ -95,111 +124,110 @@ if(isset($_POST['add']) && !empty($_POST['name'])) {
 			text-decoration: underline;
 		}
 		</style>
+		
+		<script type="text/javascript" src="js/jquery-1.12.3.min.js"></script>
+		<script type="text/javascript" src="js/jquery-ui.min.js"></script>
+		<link rel="stylesheet" href="style/jquery-ui.min.css" type="text/css">
+		
+		<script type="text/javascript" src="js/hoverIntent.js"></script>
+		<script type="text/javascript" src="js/functions.js"></script>
 	</head>
 	<body>
 		<?php
 		$files = array();
 		
-		
-		// Retriev all the php files of the created tournaments
+		// Retrieve all the php files of the created tournaments
 		foreach (glob("tournaments/*_2*.php") as $file) {
 			$file = basename($file, ".php");
 			
 			array_push($files, $file);
 		}
-		?>
-		<div id="wrapper">
-			<div id="content">
-				<h1>Toernooien</h1>
-				<?php
-				if(isset($succes)) {
-					echo $succes;
-				} elseif(isset($error)) {
-					echo $error;
+		
+		echo "<div id='wrapper'>\n";
+		echo "<div id='content'>\n";
+		echo "<h1>Toernooien</h1>\n";
+		
+		echo "<div id='response'>\n";
+		if(isset($response)) {
+			echo $response;
+		} elseif(isset($_GET['msg']) && $_GET['msg'] == "sf_add") {
+			echo "<div id='succes'>Het toernooi is toegevoegd</div>\n";
+		}
+		echo "</div>\n";
+		
+		// Initialize the form to add a new tournament
+		echo "<form name='add_tournament' method='post' action='".$_SERVER['PHP_SELF']."'>\n";
+		echo "<table>\n";
+		echo "<tr>\n";
+		echo "<td width='210'>\n";
+		
+		echo "<input type='text' name='name' value='' placeholder='Naam'>\n";
+		
+		echo "</td>\n";
+		echo "<td width='90'>\n";
+		echo "<select name='year'>\n";
+		echo "<option value='".date("Y")."'>".date("Y")."</option>\n";
+		echo "<option value='".(date("Y")+1)."'>".(date("Y")+1)."</option>\n";
+		echo "</select>\n";
+		echo "</td>\n";
+		echo "<td>\n";
+		echo "<input type='submit' name='add' value='Nieuw toernooi'>\n";
+		echo "</td>\n";
+		echo "</tr>\n";
+		echo "</table>\n";
+		echo "</form>\n";
+		
+		// If there are tournaments, as a file, display the table
+		if(sizeof($files) > 0) {
+			echo "<table>\n";
+			echo "<tr>\n";
+			echo "<th width='150'>Jaar</th>\n";
+			echo "<th width='200'>1e Plaats</th>\n";
+			echo "<th width='200'>2e Plaats</th>\n";
+			echo "<th width='75'>&nbsp;</th>\n";
+			echo "<th width='75'>&nbsp;</th>\n";
+			echo "</tr>\n";
+			
+			// For each file in the array containing files, display a table row
+			foreach($files AS $key => $filename) {
+				echo "<tr>\n";
+				echo "<td><a href='tournaments/".$filename.".php'>".str_replace("_", " ", $filename)."</a></td>\n";
+				
+				$results = mysqli_query($connect, "SELECT first, second FROM tournament_results WHERE filename = '".$filename."'");
+				$result = mysqli_fetch_assoc($results);
+				
+				if(mysqli_num_rows($results) == 0) {
+					$first = "-";
+					$second = "-";
+				} else {
+					$first = $result['first'];
+					$second = $result['second'];
 				}
 				
-				if(sizeof($files) > 0) {
-					?>
-					<table border="0" cellpadding="0" cellspacing="0">
-						<tr>
-							<th width="150">
-								Jaar
-							</th>
-							<th width="200">
-								1e Plaats
-							</th>
-							<th width="200">
-								2e Plaats
-							</th>
-							<th width="75">
-								&nbsp;
-							</th>
-							<th width="75">
-								&nbsp;
-							</th>
-						</tr>
-						<?php
-						foreach($files AS $key => $filename) {
-							echo "<tr>\n";
-							echo "<td><a href='tournaments/".$filename.".php'>".str_replace("_", " ", $filename)."</a></td>\n";
-							
-							$results = mysqli_query($connect, "SELECT first, second FROM tournament_results WHERE filename = '".$filename."'");
-							$result = mysqli_fetch_assoc($results);
-							
-							if(mysqli_num_rows($results) == 0) {
-								$first = "-";
-								$second = "-";
-							} else {
-								$first = $result['first'];
-								$second = $result['second'];
-							}
-							
-							echo "<td>".$first."</td>\n";
-							echo "<td>".$second."</td>\n";
-							echo "<td>\n";
-							echo "<form name='clear' method='post' action='".$_SERVER['PHP_SELF']."'>\n";
-							echo "<input type='hidden' name='filename' value='".$filename."'>\n";
-							echo "<input type='submit' name='clear' value='Legen'>\n";
-							echo "</form>\n";
-							echo "</td>\n";
-							echo "<td>\n";
-							echo "<form name='delete' method='post' action='".$_SERVER['PHP_SELF']."'>\n";
-							echo "<input type='hidden' name='filename' value='".$filename."'>\n";
-							echo "<input type='submit' name='delete' value='Verwijderen'>\n";
-							echo "</form>\n";
-							echo "</td>\n";
-							echo "</tr>\n";
-							
-							echo "</form>\n";
-						}
-						?>
-					</table>
-				<?php
-				} else {
-					echo "Er staan nog geen toernooien in het systeem.";
-				}
-				
-				echo "<br><br>\n";
-				echo "<form name='add' method='post' action='".$_SERVER['PHP_SELF']."'>\n";
-				echo "<table border='0' cellpadding='0' cellspacing='0'>\n";
-				echo "<tr><td width='210'>\n";
-				if(isset($_POST['submit'])) {
-					echo "<input type='text' name='name' value='".$_POST['name']."' placeholder='Naam'>\n";
-				} else {
-					echo "<input type='text' name='name' value='' placeholder='Naam'>\n";
-				}
-				echo "</td><td width='90'>\n";
-				echo "<select name='year'>\n";
-				echo "<option value='".date("Y")."'>".date("Y")."</option>\n";
-				echo "<option value='".(date("Y")+1)."'>".(date("Y")+1)."</option>\n";
-				echo "</select>\n";
-				echo "</td><td>\n";
-				echo "<input type='submit' name='add' value='Nieuw toernooi'>\n";
-				echo "</td></tr>\n";
-				echo "</table>\n";
+				echo "<td>".$first."</td>\n";
+				echo "<td>".$second."</td>\n";
+				echo "<td>\n";
+				echo "<form name='clear' method='post' action='".$_SERVER['PHP_SELF']."'>\n";
+				echo "<input type='hidden' name='filename' value='".$filename."'>\n";
+				echo "<input type='submit' name='clear' onclick='return confirm(\"Weet u zeker dat u de gegevens van dit toernooi wilt verwijderen?\")' value='Legen'>\n";
 				echo "</form>\n";
-				?>
-			</div>
-		</div>
+				echo "</td>\n";
+				echo "<td>\n";
+				echo "<form name='delete' method='post' action='".$_SERVER['PHP_SELF']."'>\n";
+				echo "<input type='hidden' name='filename' value='".$filename."'>\n";
+				echo "<input type='submit' name='delete' onclick='return confirm(\"Weet u zeker dat u dit toernooi wilt verwijderen?\")' value='Verwijderen'>\n";
+				echo "</form>\n";
+				echo "</td>\n";
+				echo "</tr>\n";
+			}
+			
+			echo "</table>\n";
+		} else {
+			echo "Er staan nog geen toernooien in het systeem.";
+		}
+		
+		echo "</div>\n";
+		echo "</div>\n";
+		?>
 	</body>
 </html>
